+++ packages/expo-router/src/typed-routes/index.ts
@@ -1,8 +1,8 @@
+import { EXPO_ROUTER_CTX_IGNORE } from 'expo-router/_ctx-shared';
 import fs from 'node:fs';
 import path from 'node:path';
 
 import { getTypedRoutesDeclarationFile } from './generate';
-import { EXPO_ROUTER_CTX_IGNORE } from '../../_ctx-shared';
 import { isTypedRoute } from '../matchers';
 import requireContext from '../testing-library/require-context-ponyfill';