+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoManager.kt
@@ -1,7 +1,10 @@
 package expo.modules.video
 
 import androidx.annotation.OptIn
+import androidx.media3.common.MediaItem
 import androidx.media3.common.util.UnstableApi
+import expo.modules.video.records.VideoSource
+import java.lang.ref.WeakReference
 
 // Helper class used to keep track of all existing VideoViews and VideoPlayers
 @OptIn(UnstableApi::class)
@@ -13,6 +16,9 @@ object VideoManager {
 
   // Keeps track of all existing VideoPlayers, and whether they are attached to a VideoView
   private var videoPlayersToVideoViews = mutableMapOf<VideoPlayer, MutableList<VideoView>>()
+
+  // Keeps track of all existing MediaItems and their corresponding VideoSources. Used for recognizing source of MediaItems.
+  private var mediaItemsToVideoSources = mutableMapOf<String, WeakReference<VideoSource>>()
 
   fun registerVideoView(videoView: VideoView) {
     videoViews[videoView.id] = videoView
@@ -32,6 +38,17 @@ object VideoManager {
 
   fun unregisterVideoPlayer(videoPlayer: VideoPlayer) {
     videoPlayersToVideoViews.remove(videoPlayer)
+  }
+
+  fun registerVideoSourceToMediaItem(mediaItem: MediaItem, videoSource: VideoSource) {
+    mediaItemsToVideoSources[mediaItem.mediaId] = WeakReference(videoSource)
+  }
+
+  fun getVideoSourceFromMediaItem(mediaItem: MediaItem?): VideoSource? {
+    if (mediaItem == null) {
+      return null
+    }
+    return mediaItemsToVideoSources[mediaItem.mediaId]?.get()
   }
 
   fun onVideoPlayerAttachedToView(videoPlayer: VideoPlayer, videoView: VideoView) {
@@ -60,7 +77,10 @@ object VideoManager {
 
   fun onAppBackgrounded() {
     for (videoView in videoViews.values) {
-      if (videoView.videoPlayer?.staysActiveInBackground == false) {
+      if (videoView.videoPlayer?.staysActiveInBackground == false &&
+        !videoView.willEnterPiP &&
+        !videoView.isInFullscreen
+      ) {
         videoView.videoPlayer?.player?.pause()
       }
     }