+++ packages/expo-video/src/index.ts
@@ -2,4 +2,9 @@ import Video from './NativeVideoModule';
 
 export { VideoView, useVideoPlayer, isPictureInPictureSupported } from './VideoView';
 export { Video };
-export { VideoSource } from './VideoView.types';
+export {
+  VideoSource,
+  VideoPlayerEvents,
+  VideoPlayerStatus,
+  VideoContentFit,
+} from './VideoView.types';